api/requirements.txt
flask
flask-cors
python-dotenv
pyodbc
Flask-JWT-Extended
axios
passlib[bcrypt]
pyotp 
qrcode[pil]